♻️ Refactor :  api 합성 및 import 수정
