+++ src/features/calendar/editSchedule/hooks/categoryQueries.ts
@@ -6,6 +6,8 @@ const useCategoryQuery = () => {
   const { data, isLoading, error } = useQuery<CategoryResponse>({
     queryKey: ['schedules', 'categories'],
     queryFn: getCategories,
+    staleTime: 10 * 60 * 1000,
+    gcTime: 30 * 60 * 1000,
   });
 
   return { data, isLoading, error };
