+++ src/screens/Calendar/CalendarSchedule/index.tsx
@@ -6,7 +6,7 @@ import { ScrollView } from 'react-native';
 import ScheduleDurationPicker from '@/features/calendar/editSchedule/components/ScheduleDurationPicker';
 import ScheduleHeaderInput from '@/features/calendar/editSchedule/components/ScheduleHeaderInput';
 import ScheduleMemoInput from '@/features/calendar/editSchedule/components/ScheduleMemoInput';
-import useScheduleEditQueries from '@/features/calendar/editSchedule/hooks/useScheduleEditQueries';
+import useScheduleEditQueries from '@/features/calendar/editSchedule/hooks/useScheduleSaveQueries';
 import CalendarDetailHeader from '@/features/calendar/editSchedule/layouts/CalendarDetailHeader';
 import useCalendarScheduleStore from '@/features/calendar/editSchedule/stores/useCalendarScheduleStore';
 import { ScheduleEditForm } from '@/features/calendar/editSchedule/types';
